requirements.txt
streamlit
numpy
pandas
matplotlib
seaborn
scikit-learn
fpdf
google-generativeai
python-dotenv
streamlit-authenticator

